refactor(tests): extract shared test fixtures to test_helpers module (CQ-13)
Deduplicate setup_store() (7 copies) and mock_embedding() (3 copies) into
src/test_helpers.rs. Removes ~95 lines of identical boilerplate and cleans
up unused imports exposed by the consolidation.
