+++ src/search.rs
@@ -1505,30 +1505,10 @@ mod tests {
     // ===== search_filtered integration tests (TC4) =====
 
     mod search_filtered_tests {
-        use crate::embedder::Embedding;
         use crate::parser::{ChunkType, Language};
-        use crate::store::helpers::{ModelInfo, SearchFilter};
-        use crate::store::Store;
+        use crate::store::helpers::SearchFilter;
+        use crate::test_helpers::{mock_embedding, setup_store};
         use std::path::PathBuf;
-
-        fn setup_store() -> (Store, tempfile::TempDir) {
-            let dir = tempfile::TempDir::new().unwrap();
-            let db_path = dir.path().join("index.db");
-            let store = Store::open(&db_path).unwrap();
-            store.init(&ModelInfo::default()).unwrap();
-            (store, dir)
-        }
-
-        fn mock_embedding(seed: f32) -> Embedding {
-            let mut v = vec![seed; 768];
-            let norm: f32 = v.iter().map(|x| x * x).sum::<f32>().sqrt();
-            if norm > 0.0 {
-                for x in &mut v {
-                    *x /= norm;
-                }
-            }
-            Embedding::new(v)
-        }
 
         fn make_chunk(
             name: &str,
