+++ src/store/types.rs
@@ -513,16 +513,8 @@ impl Store {
 mod tests {
     use super::*;
     use crate::parser::{TypeEdgeKind, TypeRef};
-    use crate::store::helpers::ModelInfo;
+    use crate::test_helpers::setup_store;
     use std::path::Path;
-
-    fn setup_store() -> (Store, tempfile::TempDir) {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("index.db");
-        let store = Store::open(&db_path).unwrap();
-        store.init(&ModelInfo::default()).unwrap();
-        (store, dir)
-    }
 
     /// Insert a minimal chunk into the store for testing type edges.
     fn insert_test_chunk(store: &Store, id: &str, name: &str, file: &str) {
